bring hcl-edit modules into rough structure
