+++ crates/hcl-edit/src/encode/expr.rs
@@ -1,0 +1,268 @@
+use super::{
+    encode_decorated, encode_escaped, Encode, EncodeDecorated, EncodeState, BOTH_SPACE_DECOR,
+    LEADING_SPACE_DECOR, NO_DECOR, TRAILING_SPACE_DECOR,
+};
+use crate::expr::{
+    Array, BinaryOp, Conditional, Expression, ForCond, ForExpr, ForIntro, FuncCall, FuncSig,
+    Object, ObjectItem, ObjectKey, ObjectKeyValueSeparator, ObjectValueTerminator, Traversal,
+    TraversalOperator, UnaryOp,
+};
+use std::fmt::{self, Write};
+
+impl EncodeDecorated for Expression {
+    fn encode_decorated(&self, buf: &mut EncodeState, default_decor: (&str, &str)) -> fmt::Result {
+        match self {
+            Expression::Null(v) => {
+                encode_decorated(v, buf, default_decor, |buf| buf.write_str("null"))
+            }
+            Expression::Bool(v) => {
+                encode_decorated(v, buf, default_decor, |buf| write!(buf, "{}", v.as_ref()))
+            }
+            Expression::Number(v) => {
+                encode_decorated(v, buf, default_decor, |buf| write!(buf, "{}", v.as_ref()))
+            }
+            Expression::String(v) => encode_decorated(v, buf, default_decor, |buf| {
+                buf.write_char('"')?;
+                encode_escaped(buf, v)?;
+                buf.write_char('"')
+            }),
+            Expression::Array(v) => v.encode_decorated(buf, default_decor),
+            Expression::Object(v) => v.encode_decorated(buf, default_decor),
+            Expression::Template(v) => v.encode_decorated(buf, default_decor),
+            Expression::HeredocTemplate(v) => v.encode_decorated(buf, default_decor),
+            Expression::Parenthesis(v) => encode_decorated(&**v, buf, default_decor, |buf| {
+                buf.write_char('(')?;
+                v.as_ref().encode_decorated(buf, NO_DECOR)?;
+                buf.write_char(')')
+            }),
+            Expression::Variable(v) => v.encode_decorated(buf, default_decor),
+            Expression::ForExpr(v) => v.encode_decorated(buf, default_decor),
+            Expression::Conditional(v) => v.encode_decorated(buf, default_decor),
+            Expression::FuncCall(v) => v.encode_decorated(buf, default_decor),
+            Expression::UnaryOp(v) => v.encode_decorated(buf, default_decor),
+            Expression::BinaryOp(v) => v.encode_decorated(buf, default_decor),
+            Expression::Traversal(v) => v.encode_decorated(buf, default_decor),
+        }
+    }
+}
+
+impl Encode for Array {
+    fn encode(&self, buf: &mut EncodeState) -> fmt::Result {
+        buf.write_char('[')?;
+
+        for (i, value) in self.values().iter().enumerate() {
+            let value_decor = if i == 0 {
+                NO_DECOR
+            } else {
+                buf.write_char(',')?;
+                LEADING_SPACE_DECOR
+            };
+            value.encode_decorated(buf, value_decor)?;
+        }
+
+        if self.trailing_comma() {
+            buf.write_char(',')?;
+        }
+
+        self.trailing().encode_with_default(buf, "")?;
+        buf.write_char(']')
+    }
+}
+
+impl Encode for Object {
+    fn encode(&self, buf: &mut EncodeState) -> fmt::Result {
+        buf.write_char('{')?;
+
+        for item in self.items().iter() {
+            item.encode(buf)?;
+        }
+
+        self.trailing().encode_with_default(buf, "")?;
+        buf.write_char('}')
+    }
+}
+
+impl Encode for ObjectItem {
+    fn encode(&self, buf: &mut EncodeState) -> fmt::Result {
+        self.key().encode_decorated(buf, TRAILING_SPACE_DECOR)?;
+
+        match self.key_value_separator() {
+            ObjectKeyValueSeparator::Colon => buf.write_char(':')?,
+            ObjectKeyValueSeparator::Equals => buf.write_char('=')?,
+        }
+
+        self.value().encode_decorated(buf, LEADING_SPACE_DECOR)?;
+
+        match self.value_terminator() {
+            ObjectValueTerminator::Comma => buf.write_char(',')?,
+            ObjectValueTerminator::Newline => buf.write_char('\n')?,
+            ObjectValueTerminator::None => {}
+        }
+
+        Ok(())
+    }
+}
+
+impl EncodeDecorated for ObjectKey {
+    fn encode_decorated(&self, buf: &mut EncodeState, default_decor: (&str, &str)) -> fmt::Result {
+        match self {
+            ObjectKey::Identifier(ident) => ident.encode_decorated(buf, default_decor),
+            ObjectKey::Expression(expr) => expr.encode_decorated(buf, default_decor),
+        }
+    }
+}
+
+impl Encode for ForExpr {
+    fn encode(&self, buf: &mut EncodeState) -> fmt::Result {
+        if let Some(key_expr) = self.key_expr() {
+            // object expr
+            buf.write_char('{')?;
+            self.intro().encode_decorated(buf, NO_DECOR)?;
+            key_expr.encode_decorated(buf, BOTH_SPACE_DECOR)?;
+            buf.write_str("=>")?;
+            self.value_expr()
+                .encode_decorated(buf, LEADING_SPACE_DECOR)?;
+            if self.grouping() {
+                buf.write_str("...")?;
+            }
+            if let Some(cond) = self.cond() {
+                cond.encode_decorated(buf, LEADING_SPACE_DECOR)?;
+            }
+            buf.write_char('}')
+        } else {
+            // list expr
+            buf.write_char('[')?;
+            self.intro().encode_decorated(buf, NO_DECOR)?;
+            self.value_expr()
+                .encode_decorated(buf, LEADING_SPACE_DECOR)?;
+            if let Some(cond) = self.cond() {
+                cond.encode_decorated(buf, LEADING_SPACE_DECOR)?;
+            }
+            buf.write_char(']')
+        }
+    }
+}
+
+impl Encode for ForIntro {
+    fn encode(&self, buf: &mut EncodeState) -> fmt::Result {
+        buf.write_str("for")?;
+        if let Some(key_var) = self.key_var() {
+            key_var.encode_decorated(buf, LEADING_SPACE_DECOR)?;
+            buf.write_char(',')?;
+        }
+        self.value_var()
+            .encode_decorated(buf, LEADING_SPACE_DECOR)?;
+        buf.write_str("in")?;
+        self.collection_expr()
+            .encode_decorated(buf, BOTH_SPACE_DECOR)?;
+        buf.write_char(':')
+    }
+}
+
+impl Encode for ForCond {
+    fn encode(&self, buf: &mut EncodeState) -> fmt::Result {
+        buf.write_str("if")?;
+        self.expr().encode_decorated(buf, LEADING_SPACE_DECOR)
+    }
+}
+
+impl Encode for Conditional {
+    fn encode(&self, buf: &mut EncodeState) -> fmt::Result {
+        self.cond_expr()
+            .encode_decorated(buf, TRAILING_SPACE_DECOR)?;
+        buf.write_char('?')?;
+        self.true_expr().encode_decorated(buf, BOTH_SPACE_DECOR)?;
+        buf.write_char(':')?;
+        self.false_expr().encode_decorated(buf, LEADING_SPACE_DECOR)
+    }
+}
+
+impl Encode for FuncCall {
+    fn encode(&self, buf: &mut EncodeState) -> fmt::Result {
+        self.name().encode_decorated(buf, NO_DECOR)?;
+        self.signature().encode_decorated(buf, NO_DECOR)
+    }
+}
+
+impl Encode for FuncSig {
+    fn encode(&self, buf: &mut EncodeState) -> fmt::Result {
+        buf.write_char('(')?;
+
+        for (i, arg) in self.args().iter().enumerate() {
+            let arg_decor = if i == 0 {
+                NO_DECOR
+            } else {
+                buf.write_char(',')?;
+                LEADING_SPACE_DECOR
+            };
+
+            arg.encode_decorated(buf, arg_decor)?;
+        }
+
+        if !self.args().is_empty() {
+            if self.expand_final() {
+                buf.write_str("...")?;
+            } else if self.trailing_comma() {
+                buf.write_char(',')?;
+            }
+        }
+
+        self.trailing().encode_with_default(buf, "")?;
+        buf.write_char(')')
+    }
+}
+
+impl Encode for UnaryOp {
+    fn encode(&self, buf: &mut EncodeState) -> fmt::Result {
+        buf.write_str(self.operator().as_str())?;
+        self.expr().encode_decorated(buf, NO_DECOR)
+    }
+}
+
+impl Encode for BinaryOp {
+    fn encode(&self, buf: &mut EncodeState) -> fmt::Result {
+        self.lhs_expr()
+            .encode_decorated(buf, TRAILING_SPACE_DECOR)?;
+        buf.write_str(self.operator().as_str())?;
+        self.rhs_expr().encode_decorated(buf, LEADING_SPACE_DECOR)
+    }
+}
+
+impl Encode for Traversal {
+    fn encode(&self, buf: &mut EncodeState) -> fmt::Result {
+        self.expr().encode_decorated(buf, NO_DECOR)?;
+
+        for operator in self.operators().iter() {
+            operator.encode_decorated(buf, NO_DECOR)?;
+        }
+
+        Ok(())
+    }
+}
+
+impl Encode for TraversalOperator {
+    fn encode(&self, buf: &mut EncodeState) -> fmt::Result {
+        match self {
+            TraversalOperator::FullSplat(_) | TraversalOperator::Index(_) => buf.write_char('[')?,
+            _other => buf.write_char('.')?,
+        }
+
+        match self {
+            TraversalOperator::AttrSplat(splat) | TraversalOperator::FullSplat(splat) => {
+                encode_decorated(splat, buf, NO_DECOR, |buf| buf.write_char('*'))?;
+            }
+            TraversalOperator::GetAttr(ident) => ident.encode_decorated(buf, NO_DECOR)?,
+            TraversalOperator::Index(expr) => expr.encode_decorated(buf, NO_DECOR)?,
+            TraversalOperator::LegacyIndex(index) => {
+                encode_decorated(index, buf, NO_DECOR, |buf| {
+                    write!(buf, "{}", index.as_ref())
+                })?;
+            }
+        }
+
+        match self {
+            TraversalOperator::FullSplat(_) | TraversalOperator::Index(_) => buf.write_char(']'),
+            _other => Ok(()),
+        }
+    }
+}
